Remove a few warnings

+++ apu/apu.cpp
@@ -342,7 +342,7 @@ void S9xResetAPU(void)
     spc::remainder = 0;
 
     SNES::cpu.reset();
-    SNES::cpu.frequency = Settings.PAL ? PAL_MASTER_CLOCK : NTSC_MASTER_CLOCK;
+    SNES::cpu.frequency = (int)(Settings.PAL ? PAL_MASTER_CLOCK : NTSC_MASTER_CLOCK);
     SNES::smp.power();
     SNES::dsp.power();
     SNES::dsp.spc_dsp.set_spc_snapshot_callback(SPCSnapshotCallback);
